requirements.txt
Django==4.2.7
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
djangorestframework==3.14.0
Faker==20.1.0
factory-boy==3.3.0
pypdf==3.17.4
djangorestframework-simplejwt==5.3.1